# Keep minimal available fd to accelerate fd allocation