@DarcySail 2019-09-09
Here are some test results.

before:
```
2019-09-09 04:13:52
Running /root/alloc_fd_benchmark
Run on (96 X 2700.04 MHz CPU s)
***WARNING*** Library was built as DEBUG. Timings may be affected.
----------------------------------------------------------
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
BM_open_serial      156868815 ns  157500000 ns          4
BM_open_zigzag      155079901 ns  155000000 ns          4
BM_open_one_by_one   56800346 ns   56666667 ns         12
BM_open_random       54486791 ns   54166667 ns         12
```

--------------------------------------------------------------------------------

after:
```
2019-09-09 04:18:17
Running /root/alloc_fd_benchmark
Run on (96 X 2700.04 MHz CPU s)
***WARNING*** Library was built as DEBUG. Timings may be affected.
----------------------------------------------------------
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
BM_open_serial       61121464 ns   62000000 ns         10
BM_open_zigzag       63065171 ns   64000000 ns         10
BM_open_one_by_one   56885521 ns   56666667 ns         12
BM_open_random       32251631 ns   32380952 ns         21
```