@DarcySail 2019-09-24
> (20:34:59) ERROR: /tmpfs/src/github/repo/test/syscalls/linux/BUILD:3479:1: no such package '@com_google_benchmark//': The repository '@com_google_benchmark' could not be resolved and referenced by '//test/syscalls/linux:alloc_fd_benchmark'

Fixed. Add corresponding reference in `WORKSPACE` file.