@DarcySail 2019-09-25
Hi @avagin and @amscanne , would you please give me more information about the compilation error?
Since `bazel test //...` will nativelly fail several test cases on my machine, I'm unable to locate where the problem is.