@DarcySail 2019-09-29
> ```
> (17:11:14) ERROR: /tmpfs/src/github/repo/test/syscalls/linux/BUILD:3531:1: Couldn't build file test/syscalls/linux/_objs/alloc_fd_benchmark/alloc_fd_benchmark.pic.o: C++ compilation of rule '//test/syscalls/linux:alloc_fd_benchmark' failed (Exit 1)
> 76
> test/syscalls/linux/alloc_fd_benchmark.cc:45:44: error: use of undeclared identifier 'MFD_CLOEXEC'
> 77
>       fd_table[i] = memfd_create(filename, MFD_CLOEXEC);
> 78
>                                            ^
> 79
> test/syscalls/linux/alloc_fd_benchmark.cc:58:44: error: use of undeclared identifier 'MFD_CLOEXEC'
> 80
>       fd_table[i] = memfd_create(filename, MFD_CLOEXEC);
> 81
>                                            ^
> 82
> test/syscalls/linux/alloc_fd_benchmark.cc:71:44: error: use of undeclared identifier 'MFD_CLOEXEC'
> 83
>       fd_table[i] = memfd_create(filename, MFD_CLOEXEC);
> 84
>                                            ^
> 85
> test/syscalls/linux/alloc_fd_benchmark.cc:86:42: error: use of undeclared identifier 'MFD_CLOEXEC'
> 86
>     fd_table[i] = memfd_create(filename, MFD_CLOEXEC);
> 87
>                                          ^
> 88
> test/syscalls/linux/alloc_fd_benchmark.cc:98:44: error: use of undeclared identifier 'MFD_CLOEXEC'
> 89
>       fd_table[j] = memfd_create(filename, MFD_CLOEXEC);
> 90
>                                            ^
> 91
> 5 errors generated.
> ```

Thanks for your feedback. In fact, I have already written a new patch to solve this compilation problem, by mannually calling `memfd_create` without glibc(It seems the glibc version on compiling machine is too old to support `memfd_create` syscall). But now, the new patch is useless since we decide not to use `memfd_create`.