@avagin 2019-10-28
https://github.com/google/gvisor/commit/fbe6b50d5628bc3d522f87eee2abcc5a923df420